Construction mode, input handling changed

+++ core/src/com/tint/hospital/render/MultipleAnimationObject.java
@@ -59,10 +59,10 @@ public class MultipleAnimationObject implements RenderObject {
 	}
 
 	@Override
-	public void setSize(float width, float height) {
+	public void setSize(int width, int height) {
 		for(TextureRegion frame : currentAnimation.getKeyFrames()) {
-			frame.setRegionWidth((int) width);
-			frame.setRegionHeight((int) height); 
+			frame.setRegionWidth(width);
+			frame.setRegionHeight(height);
 		}
 	}
 }
